Move watchdhelpers to config and rename to watchd
