+++ internal/config/watchd/testdata/golden/config_dirs_is_an_array
