+++ internal/config/watchd/testdata/golden/config_dirs_is_not_an_array
