+++ internal/config/watchd/testdata/golden/no_dirs_in_config_file
