+++ internal/config/watchd/testdata/golden/with_default_config_path
