+++ internal/config/watchd/testdata/golden/with_nested_config_path
