+++ internal/config/watchd/testdata/golden/with_relative_config_path
